+++ diive/pkgs/flux/criticaldays.py
@@ -13,13 +13,12 @@ from matplotlib.legend_handler import HandlerTuple
 from pandas import DataFrame
 
 import diive.core.plotting.styles.LightTheme as theme
-from diive.core.dfun.fits import BinFitterCP
 from diive.core.plotting import plotfuncs
-from diive.core.plotting.fitplot import fitplot
 from diive.core.plotting.plotfuncs import save_fig
 from diive.core.plotting.rectangle import rectangle
 from diive.core.plotting.styles.LightTheme import COLOR_THRESHOLD, \
-    FONTSIZE_LEGEND, COLOR_THRESHOLD2, COLOR_NEE, INFOTXT_FONTSIZE
+    FONTSIZE_LEGEND, COLOR_NEE, INFOTXT_FONTSIZE, COLOR_THRESHOLD2
+from diive.pkgs.fits.binfitter import BinFitterBTS, PlotBinFitterBTS
 
 pd.set_option('display.width', 1500)
 pd.set_option('display.max_columns', 30)
@@ -32,43 +31,48 @@ class CriticalDays:
             self,
             df: DataFrame,
             x_col: str,
-            x_units: str,
             y_col: str,
-            y_units: str,
             x_agg: Literal['max'] = 'max',
             y_agg: Literal['sum'] = 'sum',
             usebins: int = 10,
-            bootstrap_runs: int = 10,
-            bootstrapping_random_state: int = None
+            n_bootstrap_runs: int = 10,
+            bootstrapping_random_state: int = None,
+            thres_from_bootstrap: Literal['max', 'median'] = 'max'
     ):
-        """Detect critical heat days from ecosystem fluxes
+        """Detect threshold in x that defines critical days in y
 
-        XXX
+        For example, investigate daily NEE sums (y) in relation to
+        daily max VPD (x).
 
         Args:
-            df:
-            x_col: Column name of x variable
-            y_col: Column name of NEE (net ecosystem exchange, ecosystem flux)
+            df: Data in half-hourly time resolution
+            x_col: Column name of x variable, e.g. 'VPD'
+            y_col: Column name of y variable, e.g. 'NEE'
+            x_agg: Daily aggregation for x, e.g. 'max' selects the daily maximum
+            y_agg: Daily aggregation for y, e.g. 'sum' selects daily sums
             usebins: XXX (default 10)
-            bootstrap_runs: Number of bootstrap runs during detection of
+            n_bootstrap_runs: Number of bootstrap runs during detection of
                 critical heat days. Must be an odd number. In case an even
                 number is given +1 is added automatically.
+            bootstrapping_random_state: Option to use fixed random state
+            thres_from_bootstrap: Which aggregate to use as threshold,
+                e.g., for bootstrapping results [2.1, 2.2, 2.1, 2.5, 2.3] 'max'
+                selects 2.5 as threshold
         """
 
         self.df = df[[x_col, y_col]].copy()
         self.x_col = x_col
-        self.x_units = x_units
         self.x_agg = x_agg
         self.y_col = y_col
-        self.y_units = y_units
         self.y_agg = y_agg
         self.usebins = usebins
         self.bootstrapping_random_state = bootstrapping_random_state
+        self.thres_from_bootstrap = thres_from_bootstrap
 
-        # Number of bootstrap runs must be odd number
-        if bootstrap_runs % 2 == 0:
-            bootstrap_runs += 1
-        self.bootstrap_runs = bootstrap_runs
+        # # todo Number of bootstrap runs must be odd number???
+        # if bootstrap_runs % 2 == 0:
+        #     bootstrap_runs += 1
+        self.n_bootstrap_runs = n_bootstrap_runs
 
         # Resample dataset
         aggs = ['mean', 'median', 'count', 'min', 'max', 'sum']
@@ -107,28 +111,31 @@ class CriticalDays:
         thres_xcrd, num_xcrds = self._detect_xcrd_threshold()
 
         # Fit to bootstrapped data, daily time resolution
-        # Stored as bootstrap runs > 0 (bts>0)
-        bts_results = self._bootstrap_fits(df_aggs=self.df_aggs,
-                                           x_col=self.x_col,
-                                           x_agg=self.x_agg,
-                                           y_col=self.y_col,
-                                           y_agg=self.y_agg,
-                                           fit_to_bins=self.usebins,
-                                           detect_zerocrossing=True)
+        fit_results, bts_fit_results = self._bootstrap_fits(df_aggs=self.df_aggs,
+                                                            x_col=self.x_col,
+                                                            x_agg=self.x_agg,
+                                                            y_col=self.y_col,
+                                                            y_agg=self.y_agg,
+                                                            fit_to_bins=self.usebins)
+
+        # Add info about zero crossings to fit results
+        fit_results = self._collect_zerocrossings(fit_results=fit_results)
+        for key, cur_fit_results in bts_fit_results.items():
+            bts_fit_results[key] = self._collect_zerocrossings(fit_results=bts_fit_results[key])
 
         # Get flux zerocrossing values (y has crossed zeroline) from bootstrap runs
-        bts_zerocrossings_df = self._bts_zerocrossings_collect(bts_results=bts_results)
+        bts_zerocrossings_df = self._bts_zerocrossings_collect(bts_fit_results=bts_fit_results)
 
         # Calc flux equilibrium points aggregates from bootstrap runs
         bts_zerocrossings_aggs = self._zerocrossings_aggs(bts_zerocrossings_df=bts_zerocrossings_df)
 
         # Threshold for Critical Heat Days (CRDs)
-        # defined as the linecrossing MEDIAN x (e.g. VPD) from bootstrap runs
-        thres_crd = bts_zerocrossings_aggs['x_median']
-        # thres_crd = bts_zerocrossings_aggs['x_max']
+        # defined as the linecrossing e.g. MAX x (e.g. VPD) from bootstrap runs
+        # thres_crd = bts_zerocrossings_aggs['x_median']
+        thres_crd = bts_zerocrossings_aggs[f"x_{self.thres_from_bootstrap}"]  # e.g. "x_max"
 
         # Collect days above or equal to CRD threshold
-        df_aggs_crds = self.df_aggs.loc[self.df_aggs[self.x_col]['max'] >= thres_crd, :].copy()
+        df_aggs_crds = self.df_aggs.loc[self.df_aggs[self.x_col][self.x_agg] >= thres_crd, :].copy()
 
         # Number of days above CRD threshold
         num_crds = len(df_aggs_crds)
@@ -138,7 +145,7 @@ class CriticalDays:
         # of days below of equal to CRD threshold.
         # For example: if 10 CRDs were found, nCRDs are the 10 days closest
         # to the CRD threshold (below or equal to the threshold).
-        sortby_col = (self.x_col, 'max')
+        sortby_col = (self.x_col, self.x_agg)
         ncrds_start_ix = num_crds
         ncrds_end_ix = num_crds * 2
         df_aggs_ncrds = self.df_aggs \
@@ -146,8 +153,9 @@ class CriticalDays:
                             .iloc[ncrds_start_ix:ncrds_end_ix]
 
         # Threshold for nCRDs
-        # The lower threshold is the minimum of found x maxima
-        thres_ncrds_lower = df_aggs_ncrds[self.x_col]['max'].min()
+        # The lower threshold is e.g. the minimum of found x maxima, depending
+        # on the param "thres_from_bootstrap"
+        thres_ncrds_lower = df_aggs_ncrds[self.x_col][self.x_agg].min()
         thres_ncrds_upper = thres_crd
 
         # Number of days above nCRD threshold and below or equal CRD threshold
@@ -155,7 +163,7 @@ class CriticalDays:
 
         # Collect results
         self._results_threshold_detection = dict(
-            bts_results=bts_results,
+            fit_results=fit_results,  # Non-boostrapped fit results
             bts_zerocrossings_df=bts_zerocrossings_df,
             bts_zerocrossings_aggs=bts_zerocrossings_aggs,
             thres_crd=thres_crd,
@@ -170,105 +178,86 @@ class CriticalDays:
         )
 
     def plot_crd_detection_results(self, ax,
+                                   x_units: str,
+                                   y_units: str,
                                    showfit: bool = True,
                                    highlight_year: int = None,
                                    showline_crd: bool = True,
                                    showrange_crd: bool = True,
                                    showline_xcrd: bool = True):
-        """Plot results from critical heat days threshold detection"""
+        """Plot results from critical days threshold detection"""
 
-        # reload(diive.core.plotting.styles.LightTheme)
-        # import diive.core.plotting.styles.LightTheme
-
-        bts_results = self.results_threshold_detection['bts_results'][0]  # 0 means non-bootstrapped data
+        # bts_results = self.results_threshold_detection['bts_results'][0]  # 0 means non-bootstrapped data
+        fit_results = self.results_threshold_detection['fit_results']
         bts_zerocrossings_aggs = self.results_threshold_detection['bts_zerocrossings_aggs']
 
-        # y
-        line_xy_y = line_fit_y = line_fit_ci_y = line_fit_pb_y = line_highlight = None
-        _numvals_per_bin = bts_results[self.y_col]['numvals_per_bin']
-        flux_bts_results = bts_results[self.y_col]
-        line_xy_y, line_fit_y, line_fit_ci_y, line_fit_pb_y, line_highlight = \
-            fitplot(ax=ax,
-                    label="NEE",
-                    highlight_year=highlight_year,
-                    flux_bts_results=flux_bts_results,
-                    edgecolor='#B0BEC5', color='none', color_fitline=COLOR_NEE,
-                    showfit=showfit, show_prediction_interval=True)
+        self.ax, line_xy, line_highlight, line_fit, line_fit_ci, line_fit_pb = \
+            PlotBinFitterBTS(fit_results=fit_results,
+                             ax=ax, label="NEE",
+                             highlight_year=highlight_year, edgecolor='#B0BEC5',
+                             color='none', color_fitline=COLOR_NEE,
+                             showfit=showfit, show_prediction_interval=True).plot_binfitter()
 
-        # # Actual non-bootstrapped line crossing, the point where RECO = GPP
-        # line_netzeroflux = ax.scatter(bts_results['zerocrossing_vals']['x_col'],
-        #                               bts_results['zerocrossing_vals']['y_nom'],
-        #                               edgecolor='black', color='none', alpha=1, s=90, lw=2,
-        #                               label='net zero flux', zorder=99, marker='s')
-
-        # # Rectangle
-        # # Bootstrapped line crossing, the point where RECO = GPP
-        # line_bts_median_netzeroflux = ax.scatter(bts_zerocrossings_aggs['x_max'],
-        #                                          0,
-        #                                          edgecolor=COLOR_NEE, color='none', alpha=1, s=350, lw=3,
-        #                                          label='net zero flux (bootstrapped median)', zorder=97, marker='s')
-
-        # Threshold lines
-
-        ## Vertical line showing MEDIAN CRD threshold from bootstrap runs
+        # Threshold line
+        # Vertical line showing e.g. MAX CRD threshold from bootstrap runs
         line_crd_vertical = None
         if showline_crd:
             _sub = "$_{CRD}$"
-            line_crd_vertical = ax.axvline(bts_zerocrossings_aggs['x_median'], lw=3, color=COLOR_THRESHOLD, ls='-',
-                                           zorder=99,
-                                           label=f"THR{_sub}, VPD = {bts_zerocrossings_aggs['x_median']:.2f} {self.x_units}")
-            # # CRDs
+            thres_crd = self.results_threshold_detection['thres_crd']
+            line_crd_vertical = ax.axvline(thres_crd, lw=3,
+                                           color=COLOR_THRESHOLD, ls='-', zorder=99,
+                                           label=f"THR{_sub}, VPD = {thres_crd:.2f} {x_units}")
+            # Text CRDs
             num_crds = len(self.results_threshold_detection['df_aggs_crds'])
-            # area_crd = ax.fill_between([bts_zerocrossings_aggs['x_max'],
-            #                             flux_bts_results['fit_df']['fit_x'].max()],
-            #                            0, 1,
-            #                            color='#BA68C8', alpha=0.1, transform=ax.get_xaxis_transform(),
-            #                            label=f"CRDs ({num_crds} days)", zorder=1)
             sym_max = r'$\rightarrow$'
-            _pos = bts_zerocrossings_aggs['x_median'] + bts_zerocrossings_aggs['x_median'] * 0.02
+            _pos = thres_crd + thres_crd * 0.02
             t = ax.text(_pos, 0.1, f"{sym_max} {num_crds} critical days {sym_max}", size=INFOTXT_FONTSIZE,
                         color='black', backgroundcolor='none', transform=ax.get_xaxis_transform(),
                         alpha=1, horizontalalignment='left', verticalalignment='top', zorder=99)
             t.set_bbox(dict(facecolor=COLOR_THRESHOLD, alpha=.7, edgecolor=COLOR_THRESHOLD))
 
-        ## Rectangle bootstrap range CRD
+            # # Area CRDs
+            # area_crd = ax.fill_between([bts_zerocrossings_aggs[_thres_agg],
+            #                             fit_results['fit_df']['fit_x'].max()],
+            #                            0, 1, color='#BA68C8', alpha=0.1,
+            #                            transform=ax.get_xaxis_transform(),
+            #                            label=f"CRDs ({num_crds} days)", zorder=1)
+
+        # ## Rectangle bootstrap range CRD
         range_bts_netzeroflux = None
         if showrange_crd:
-            num_bootstrap_runs = len(
-                self.results_threshold_detection['bts_results'].keys()) - 1  # -1 b/c zero run is non-bootstrapped
             _sub = "$_{CRD}$"
-            range_bts_netzeroflux = rectangle(ax=ax,
-                                              rect_lower_left_x=bts_zerocrossings_aggs['x_min'],
-                                              rect_lower_left_y=0,
-                                              rect_width=bts_zerocrossings_aggs['x_max'] - bts_zerocrossings_aggs[
-                                                  'x_min'],
-                                              rect_height=1,
-                                              label=f"THR{_sub} range ({num_bootstrap_runs} bootstraps)",
-                                              color=COLOR_THRESHOLD)
+            range_bts_netzeroflux = rectangle(
+                ax=ax,
+                rect_lower_left_x=bts_zerocrossings_aggs['x_min'],
+                rect_lower_left_y=0,
+                rect_width=bts_zerocrossings_aggs['x_max'] - bts_zerocrossings_aggs['x_min'],
+                rect_height=1,
+                label=f"THR{_sub} range ({self.n_bootstrap_runs} bootstraps)",
+                color=COLOR_THRESHOLD)
 
-        ## Vertical line showing xCRD threshold
+        # ## Vertical line showing xCRD threshold
         line_xcrd_vertical = None
         if showline_xcrd:
             thres_xcrd = self.results_threshold_detection['thres_xcrd']
             num_xcrds = self.results_threshold_detection['num_xcrds']
             _sub = "$_{XCRD}$"
             line_xcrd_vertical = ax.axvline(thres_xcrd, lw=3, color=COLOR_THRESHOLD2, ls='-', zorder=99,
-                                            label=f"THR{_sub}, VPD = {thres_xcrd:.2f} {self.x_units}")
+                                            label=f"THR{_sub}, VPD = {thres_xcrd:.2f} {x_units}")
             # t = ax.text(thres_xcrd + thres_xcrd * 0.02, 0.04, f"{num_xcrd} extremely critical days",
             #             size=INFOTXT_FONTSIZE,
             #             color='black', backgroundcolor='none', transform=ax.get_xaxis_transform(),
             #             alpha=1, horizontalalignment='left', verticalalignment='top', zorder=99)
 
-
-
         # Format
         ax.axhline(0, lw=1, color='black')
-        xlabel = f"Daily maximum VPD (${self.x_units}$)"
+        xlabel = f"Daily maximum VPD (${x_units}$)"
 
-        ylabel = f"{self.y_col} (${self.y_units}$)"
+        ylabel = f"{self.y_col} (${y_units}$)"
         plotfuncs.default_format(ax=ax, txt_xlabel=xlabel, txt_ylabel=ylabel)
 
         # Custom legend
+        # This legend replaces the legend from PlotBinFitterBTS
         # Assign two of the handles to the same legend entry by putting them in a tuple
         # and using a generic handler map (which would be used for any additional
         # tuples of handles like (p1, p3)).
@@ -278,21 +267,21 @@ class CriticalDays:
                 line_crd_vertical if showline_crd else None,
                 range_bts_netzeroflux if showrange_crd else None,
                 line_xcrd_vertical if showline_xcrd else None,
-                line_xy_y,
+                line_xy,
                 line_highlight if line_highlight else None,
-                line_fit_y if showfit else None,
-                line_fit_ci_y if showfit else None,
-                line_fit_pb_y if showfit else None
+                line_fit if showfit else None,
+                line_fit_ci if showfit else None,
+                line_fit_pb if showfit else None
             ],
             [
                 line_crd_vertical.get_label() if showline_crd else None,
                 range_bts_netzeroflux.get_label() if showrange_crd else None,
                 line_xcrd_vertical.get_label() if showline_xcrd else None,
-                line_xy_y.get_label(),
+                line_xy.get_label(),
                 line_highlight.get_label() if line_highlight else None,
-                line_fit_y.get_label() if showfit else None,
-                line_fit_ci_y.get_label() if showfit else None,
-                line_fit_pb_y.get_label() if showfit else None
+                line_fit.get_label() if showfit else None,
+                line_fit_ci.get_label() if showfit else None,
+                line_fit_pb.get_label() if showfit else None
             ],
             bbox_to_anchor=(0, 1),
             frameon=False,
@@ -306,6 +295,8 @@ class CriticalDays:
         ax.text(0.06, 0.95, "(a)",
                 size=theme.AXLABELS_FONTSIZE, color='black', backgroundcolor='none', transform=ax.transAxes,
                 alpha=1, horizontalalignment='left', verticalalignment='top')
+
+        return ax
 
     def _resample_dataset(self, df: DataFrame, aggs: list, day_start_hour: int = None):
         """Resample to daily values from *day_start_hour* to *day_start_hour*"""
@@ -333,13 +324,14 @@ class CriticalDays:
 
         return zerocrossings_aggs
 
-    def _bts_zerocrossings_collect(self, bts_results):
+    def _bts_zerocrossings_collect(self, bts_fit_results) -> DataFrame:
+        """Collect all zero crossing values in DataFrame"""
         bts_linecrossings_df = pd.DataFrame()
-        for bts in range(1, self.bootstrap_runs + 1):  # +1 because 0 = non-bootstrapped run
-            _dict = bts_results[bts]['zerocrossing_vals']
+        for bts in range(0, self.n_bootstrap_runs):
+            _dict = bts_fit_results[bts]['zerocrossing_vals']
             _series = pd.Series(_dict)
             _series.name = bts
-            if bts == 1:
+            if bts == 0:
                 bts_linecrossings_df = pd.DataFrame(_series).T
             else:
                 bts_linecrossings_df = bts_linecrossings_df.append(_series.T)
@@ -351,74 +343,44 @@ class CriticalDays:
                         x_agg: str,
                         y_col: str,
                         y_agg: str,
-                        fit_to_bins: int = 10,
-                        detect_zerocrossing: bool = False) -> dict:
+                        fit_to_bins: int = 10) -> tuple[dict, dict]:
         """Bootstrap ycols and fit to x"""
 
         # Get column names in aggregated df
         x_col = (x_col, x_agg)
         y_col = (y_col, y_agg)
-        bts_results = {}
-        bts = 0
 
-        while bts < self.bootstrap_runs + 1:
-            print(f"Bootstrap run #{bts}")
-            fit_results = {}
+        fitter = BinFitterBTS(
+            df=df_aggs,
+            n_bootstraps=self.n_bootstrap_runs,
+            x_col=x_col,
+            y_col=y_col,
+            predict_max_x=self.predict_max_x,
+            predict_min_x=self.predict_min_x,
+            num_predictions=1000,
+            bins_x_num=0,
+            # bins_y_agg='mean',
+            fit_type='quadratic'
+        )
+        fitter.fit()
 
-            if bts > 0:
-                # Bootstrap data
-                bts_df = df_aggs.sample(n=int(len(df_aggs)), replace=True, random_state=self.bootstrapping_random_state)
-            else:
-                # First run (bts=0) is with measured data (not bootstrapped)
-                bts_df = df_aggs.copy()
+        fit_results = fitter.fit_results
+        bts_fit_results = fitter.bts_fit_results
 
-            try:
-                # Fit
-                fitter = BinFitterCP(df=bts_df,
-                                     x_col=x_col,
-                                     y_col=y_col,
-                                     num_predictions=1000,
-                                     predict_min_x=self.predict_min_x,
-                                     predict_max_x=self.predict_max_x,
-                                     bins_x_num=fit_to_bins,
-                                     bins_y_agg='mean',
-                                     fit_type='quadratic')
-                fitter.run()
-                cur_fit_results = fitter.get_results()
+        return fit_results, bts_fit_results
 
-                # Store fit results for current y
-                fit_results[y_col[0]] = cur_fit_results
+    def _collect_zerocrossings(self, fit_results) -> dict:
+        """Collect zero-crossing results from all fit results"""
+        zerocrossing_vals = self._detect_zerocrossing_y(fit_df=fit_results['fit_df'])
+        if isinstance(zerocrossing_vals, dict):
+            pass
+        else:
+            raise ValueError
+        fit_results['zerocrossing_vals'] = zerocrossing_vals
+        print(f"Found zero-crossing values: {fit_results['zerocrossing_vals']}")
+        return fit_results
 
-                # Zero crossing for y
-                if detect_zerocrossing:
-                    zerocrossing_vals = \
-                        self._detect_zerocrossing_y(fit_results=cur_fit_results['fit_df'])
-
-                    if isinstance(zerocrossing_vals, dict):
-                        pass
-                    else:
-                        raise ValueError
-
-                    fit_results['zerocrossing_vals'] = zerocrossing_vals
-                    print(fit_results['zerocrossing_vals'])
-
-                # import matplotlib.pyplot as plt
-                # fit_results['fit_df'][['fit_x', 'nom']].plot()
-                # plt.scatter(fit_results['fit_df']['fit_x'], fit_results['fit_df']['nom'])
-                # plt.scatter(df[x_col], df[y_col])
-                # plt.show()
-
-
-            except ValueError:
-                print(f"(!) WARNING Bootstrap run #{bts} was not successful, trying again")
-
-            # Store bootstrap results in dict
-            bts_results[bts] = fit_results
-            bts += 1
-
-        return bts_results
-
-    def _detect_zerocrossing_y(self, fit_results: dict):
+    def _detect_zerocrossing_y(self, fit_df: DataFrame):
         # kudos: https://stackoverflow.com/questions/28766692/intersection-of-two-graphs-in-python-find-the-x-value
 
         num_zerocrossings = None
@@ -426,8 +388,8 @@ class CriticalDays:
 
         # Collect predicted vals in df
         zerocrossings_df = pd.DataFrame()
-        zerocrossings_df['x_col'] = fit_results['fit_x']
-        zerocrossings_df['y_nom'] = fit_results['nom']
+        zerocrossings_df['x_col'] = fit_df['fit_x']
+        zerocrossings_df['y_nom'] = fit_df['nom']
 
         # Check values above/below zero
         _signs = np.sign(zerocrossings_df['y_nom'])
@@ -504,10 +466,10 @@ class CriticalDays:
         return agg_df
 
     def showplot_criticaldays(self,
-                     saveplot: bool = False,
-                     title: str = None,
-                     path: Path or str = None,
-                     **kwargs):
+                              saveplot: bool = False,
+                              title: str = None,
+                              path: Path or str = None,
+                              **kwargs):
         fig = plt.figure(figsize=(9, 9))
         gs = gridspec.GridSpec(1, 1)  # rows, cols
         # gs.update(wspace=.2, hspace=1, left=.1, right=.9, top=.85, bottom=.1)
@@ -520,46 +482,58 @@ class CriticalDays:
 
 
 if __name__ == '__main__':
-    pd.options.display.width = None
-    pd.options.display.max_columns = None
-    # pd.set_option('display.max_rows', 3000)
-    pd.set_option('display.max_columns', 3000)
+    # # Subset
+    # subset_cols = [vpd_col, nee_col]
+    # df = df_orig[subset_cols].copy().dropna()
+    #
+    # # Convert units
+    # df[vpd_col] = df[vpd_col].multiply(0.1)  # hPa --> kPa
+    # df[nee_col] = df[nee_col].multiply(0.0792171)  # umol CO2 m-2 s-1 --> g CO2 m-2 30min-1
+    # df = df.loc[df.index.year <= 2022]
+    # df = df.loc[(df.index.month >= 5) & (df.index.month <= 9)]
 
-    # Test data
     from diive.core.io.files import load_pickle
 
-    df_orig = load_pickle(
-        filepath=r'L:\Dropbox\luhk_work\_current\fp2022\7-14__IRGA627572__addingQCF0\CH-DAV_FP2022.1_1997-2022.08_ID20220826234456_30MIN.diive.csv.pickle')
-
-    # Settings
+    # Variables
     vpd_col = 'VPD_f'
     nee_col = 'NEE_CUT_REF_f'
-    subset_cols = [vpd_col, nee_col]
-    df = df_orig[subset_cols].copy().dropna()
+    x_col = vpd_col
+    y_col = nee_col
+
+    # Load data, using pickle for fast loading
+    source_file = r"L:\Dropbox\luhk_work\20 - CODING\21 - DIIVE\diive\__apply\co2penalty_dav\input_data\CH-DAV_FP2022.1_1997-2022.08_ID20220826234456_30MIN.diive.csv.pickle"
+    df_orig = load_pickle(filepath=source_file)
+    df_orig = df_orig.loc[df_orig.index.year <= 2022].copy()
+    # df_orig = df_orig.loc[df_orig.index.year >= 2019].copy()
+
+    # Select daytime data between May and September 1997-2021
+    maysep_dt_df = df_orig.loc[(df_orig.index.month >= 5) & (df_orig.index.month <= 9)].copy()
+
     # Convert units
-    df[vpd_col] = df[vpd_col].multiply(0.1)  # hPa --> kPa
-    df[nee_col] = df[nee_col].multiply(0.0792171)  # umol CO2 m-2 s-1 --> g CO2 m-2 30min-1
-    df = df.loc[df.index.year < 2022]
-    df = df.loc[(df.index.month >= 5) & (df.index.month <= 9)]
+    maysep_dt_df[vpd_col] = maysep_dt_df[vpd_col].multiply(0.1)  # hPa --> kPa
+    maysep_dt_df[nee_col] = maysep_dt_df[nee_col].multiply(0.0792171)  # umol CO2 m-2 s-1 --> g CO2 m-2 30min-1
+    x_units = "kPa"
+    y_units = "gCO_{2}\ m^{-2}\ 30min^{-1}"
+    xlabel = f"Half-hourly VPD ({x_units})"
+    ylabel = f"{y_col} (${y_units}$)"
 
     # Critical days
     crd = CriticalDays(
-        df=df,
+        df=maysep_dt_df,
         x_col=vpd_col,
-        x_units='kPa',
         x_agg='max',
         y_col=nee_col,
-        y_units="gCO_{2}\ m^{-2}\ d^{-1}",
         y_agg='sum',
         usebins=0,
-        bootstrap_runs=3,
-        bootstrapping_random_state=None
+        n_bootstrap_runs=99,
+        bootstrapping_random_state=None,
+        thres_from_bootstrap='max'
     )
-
     crd.detect_crd_threshold()
-    results = crd.results_threshold_detection
-    crd.plot_results(showfit=True,
-                     highlight_year=2019,
-                     showline_crd=True,
-                     showrange_crd=True,
-                     showline_xcrd=True)
+    results_threshold_detection = crd.results_threshold_detection
+    crd.showplot_criticaldays(x_units='kPa',
+                              y_units="gCO_{2}\ m^{-2}\ d^{-1}",
+                              highlight_year=2019,
+                              showline_crd=True,
+                              showrange_crd=True,
+                              showline_xcrd=False)