+++ diive/pkgs/gapfilling/randomforest_ts.py
@@ -426,7 +426,7 @@ class RandomForestTS:
         # Targets w/ full timestamp, gaps in this series will be filled
         targets_ser = pd.Series(data=df[self.target_col], index=df.index)
 
-        # Remove target from df
+        # Remove target from data
         df.drop(self.target_col, axis=1, inplace=True)
 
         # df now contains features only, keep rows with complete features
@@ -595,9 +595,9 @@ class RandomForestTS:
 
 def example():
     import matplotlib.pyplot as plt
-    from diive.core.io.files import load_pickle
     from diive.core.plotting.heatmap_datetime import HeatmapDateTime
 
+    # # Read data file
     # from diive.core.io.filereader import ReadFileType
     # loaddatafile = ReadFileType(
     #     filetype='REDDYPROC_30MIN',
@@ -609,23 +609,47 @@ def example():
     #                filename='testdata_CH-DAV_FP2021.2_2016-2020_ID20220324003457_30MIN_SUBSET.csv',
     #                data=data_df)
 
-    # Test data
-    df = load_pickle(
-        filepath=r'L:\Dropbox\luhk_work\20 - CODING\21 - DIIVE\diive\tests\testdata\testdata_CH-DAV_FP2021.2_2016-2020_ID20220324003457_30MIN_SUBSET.csv.pickle')
-    df.drop('NEE_CUT_f', axis=1, inplace=True)
+    # # Test data
+    # from diive.core.io.files import load_pickle
+    # df = load_pickle(
+    #     filepath=r'L:\Dropbox\luhk_work\20 - CODING\21 - DIIVE\diive\tests\testdata\testdata_CH-DAV_FP2021.2_2016-2020_ID20220324003457_30MIN_SUBSET.csv.pickle')
+    # df.drop('NEE_CUT_f', axis=1, inplace=True)
+
+    # Download from database
+    from dbc_influxdb import dbcInflux
+    DIRCONF = r'L:\Dropbox\luhk_work\20 - CODING\22 - POET\configs'  # Folder with configurations
+    SITE = 'ch-dav'  # Site name
+    BUCKET_PROCESSING = f"{SITE}_processing"
+    TARGET_COL = 'TA_NABEL_T1_35_1'
+    MEASUREMENTS = ['TA', 'SW', 'RH']  # Measurement name, e.g., 'TA' contains all air temperature variables
+    FIELDS = ['TA_NABEL_T1_35_1', 'SW_IN_NABEL_T1_35_1', 'RH_NABEL_T1_35_1']  # Variable name; InfluxDB stores variable names as '_field'
+    DATA_VERSION = 'meteoscreening'
+    START = '2021-10-01 00:01:00'  # Download data starting with this date
+    STOP = '2021-12-01 00:01:00'  # Download data before this date (the stop date itself is not included)
+    TIMEZONE_OFFSET_TO_UTC_HOURS = 1  # Timezone, e.g. "1" is translated to timezone "UTC+01:00" (CET, winter time)
+
+    dbc = dbcInflux(dirconf=DIRCONF)
+
+    data_simple, data_detailed, assigned_measurements = dbc.download(bucket=BUCKET_PROCESSING,
+                                                                     measurements=MEASUREMENTS,
+                                                                     fields=FIELDS,
+                                                                     start=START,
+                                                                     stop=STOP,
+                                                                     timezone_offset_to_utc_hours=TIMEZONE_OFFSET_TO_UTC_HOURS,
+                                                                     data_version=DATA_VERSION)
 
     # df = df.loc[df.index.year == 2016, :].copy()
-    df = df[['NEE_CUT_orig']].copy()
+    df = data_simple.copy()
 
     # Random forest
     rfts = RandomForestTS(df=df,
-                          target_col='NEE_CUT_orig',
+                          target_col=TARGET_COL,
                           include_timestamp_as_features=True,
-                          lagged_variants=1,
+                          lagged_variants=3,
                           use_neighbor_years=True,
-                          feature_reduction=True,
+                          feature_reduction=False,
                           verbose=1)
-    rfts.build_models(n_estimators=5,
+    rfts.build_models(n_estimators=200,
                       random_state=42,
                       min_samples_split=2,
                       min_samples_leaf=1,
@@ -633,11 +657,26 @@ def example():
     rfts.gapfill_yrs()
     gapfilled_yrs_df, yrpool_gf_results = rfts.get_gapfilled_dataset()
 
-    HeatmapDateTime(series=rfts.gapfilled_yrs_df['NEE_CUT_orig']).show()
-    HeatmapDateTime(series=rfts.gapfilled_yrs_df['NEE_CUT_orig_gfRF']).show()
+    # Gap-filled data series
+    target_col_gf = f'{TARGET_COL}_gfRF'
+    gapfilled_series = gapfilled_yrs_df[target_col_gf].copy()
 
-    gapfilled_yrs_df['.gapfilled_cumulative'].plot()
-    plt.show()
+    # Plot
+    HeatmapDateTime(series=rfts.gapfilled_yrs_df[TARGET_COL]).show()
+    HeatmapDateTime(series=rfts.gapfilled_yrs_df[target_col_gf]).show()
+    # gapfilled_yrs_df['.gapfilled_cumulative'].plot()
+    # plt.show()
+
+    # Merge gapfilled series with tags
+    data_detailed_gf_var = data_detailed[TARGET_COL].copy()  # Original data with tags
+    data_detailed_gf_var[TARGET_COL] = gapfilled_series  # Replace original series w/ gapfilled series (same name)
+    data_detailed_gf_var['data_version'] = 'gapfilled'  # Change data version to 'gapfilled'
+
+    # Upload gapfilled var to database
+    dbc.upload_singlevar(to_bucket=BUCKET_PROCESSING,
+                         to_measurement=assigned_measurements[TARGET_COL],
+                         var_df=data_detailed_gf_var,
+                         timezone_of_timestamp='UTC+01:00')
 
     print("Finished.")